Add: разбивает на отдельные компоненты
