--- src/view/films-list-view.js
@@ -1,27 +1,0 @@
-import { createElement } from '../render';
-
-const createFilmsListTemplate = () => `
-<section class="films-list">
-  <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-  <div class="films-list__container"></div>
-</section>`;
-
-export default class FilmsList {
-  #element = null;
-
-  get template() {
-    return createFilmsListTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
-  }
-}
