+++ src/view/films-list/films-list-container-view.js
@@ -1,0 +1,23 @@
+import { createElement } from '../../render';
+
+const createFilmsListContainerTemplate = () => '<div class="films-list__container"></div>';
+
+export default class FilmsListContainer {
+  #element = null;
+
+  get template() {
+    return createFilmsListContainerTemplate();
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
